pyproject.toml
[tool.poetry]
name = "python_django_diploma"
version = "0.1.0"
description = ""
authors = ["Alexandr Basso <alexandr.basso@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.0"
djangorestframework = "^3.14.0"
pillow = "^10.0.0"
django-debug-toolbar = "^4.1.0"
black = "^23.7.0"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
